Moved own classes to "classes/local" to comply to moodle standards as mentioned in Issue https://github.com/SE-Stuttgart/moodle-block_booksearch/issues/26.
