+++ classes/local/search.php
@@ -22,7 +22,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace block_booksearch\search;
+namespace block_booksearch\local;
 
 use stdClass;
